Rename foler /localization -> /l10n
