+++ src/client/l10n/locales.ts
