+++ src/client/pages/LoginApp/components/LoginForm.tsx
@@ -3,7 +3,7 @@ import { colorsV3 } from '@hedviginsurance/brand'
 import { Form, Formik, FormikHelpers } from 'formik'
 import React from 'react'
 import * as Yup from 'yup'
-import { LocaleLabel, locales } from 'localization/locales'
+import { LocaleLabel, locales } from 'l10n/locales'
 import { Button } from 'components/buttons'
 import { InputField } from 'components/inputs'
 import { LoadingDots } from 'components/LoadingDots/LoadingDots'
